# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: iHub - Marketplace Management Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/search:
    get:
      operationId: searchInventory
      tags: [inventory]
      summary: Buscar SKUs para inventário geral
      description: >-
        Lista anúncios com SKU (Full e não Full), agrupados por SKU (título, variação, estoque mandatário).
        Busca por SKU, MLB, título ou texto em variações (mesma lógica de tokens que /products).
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: >-
            Texto ou código de barras / SKU. Se omitido ou vazio, retorna todos os SKUs
            de anúncios da conta (Full e não Full).
      responses:
        "200":
          description: Resultados agrupados por SKU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventorySearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /inventory/sku/{sku}/financials:
    patch:
      operationId: patchInventorySkuFinancials
      tags: [inventory]
      summary: Salvar imposto e preço de compra por SKU
      description: >-
        Persiste percentual de imposto (0–100) e preço de compra (BRL) para o SKU no escopo do usuário.
        Exige pelo menos um anúncio (Full ou não Full) com esse SKU. Campos omitidos mantêm o valor anterior.
      security:
        - bearerAuth: []
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
          description: SKU (URL-encoded se necessário)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventorySkuFinancialsPatchRequest"
      responses:
        "200":
          description: Valores salvos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventorySkuFinancialsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /inventory/mandate-adjust:
    post:
      operationId: adjustMandateInventory
      tags: [inventory]
      summary: Ajustar estoque mandatário e espelhar nos anúncios
      description: >-
        Atualiza sku_mandate_inventory e aplica a quantidade em todos os anúncios não Full
        com o mesmo SKU nas contas do usuário.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MandateAdjustRequest"
      responses:
        "200":
          description: Resultado do espelhamento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MandateAdjustResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /inventory/sync-sku:
    post:
      operationId: syncSkuStock
      tags: [inventory]
      summary: Forçar re-sincronização de estoque por SKU
      description: >-
        Lê o estoque atual no ML do anúncio fonte (o indicado por sourceProductId ou o primeiro ativo)
        e propaga o valor para todos os anúncios não-Full com o mesmo SKU nas contas do usuário.
        Anúncios Full são completamente ignorados.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkuSyncRequest"
      responses:
        "200":
          description: Resultado da sincronização
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkuSyncResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Orders
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    InventorySearchItem:
      type: object
      properties:
        sku:
          type: string
        mandateQuantity:
          type: integer
          nullable: true
          description: Valor em sku_mandate_inventory, se existir
        thumbnail:
          type: string
          nullable: true
        titleShort:
          type: string
        variationLabel:
          type: string
          nullable: true
          description: Cor / tamanho etc. quando há variações
        currentStock:
          type: integer
          description: Estoque no anúncio representativo (banco)
        representativeProductId:
          type: string
        listingCount:
          type: integer
          description: Quantidade total de anúncios (Full e não Full) com este SKU
        nonFullListingCount:
          type: integer
          description: Quantidade de anúncios não Full com este SKU (0 = só Full)
        taxPercent:
          type: number
          nullable: true
          description: Percentual de imposto (0–100) salvo pelo usuário
        purchasePrice:
          type: number
          nullable: true
          description: Preço de compra de referência (BRL)
      required:
        - sku
        - titleShort
        - currentStock
        - representativeProductId
        - listingCount
        - nonFullListingCount
    InventorySkuFinancialsPatchRequest:
      type: object
      properties:
        taxPercent:
          type: number
          nullable: true
          description: Percentual de imposto (0–100); null remove o valor
        purchasePrice:
          type: number
          nullable: true
          description: Preço de compra em BRL; null remove o valor
      description: Pelo menos uma propriedade deve ser enviada.
    InventorySkuFinancialsResponse:
      type: object
      properties:
        sku:
          type: string
        taxPercent:
          type: number
          nullable: true
        purchasePrice:
          type: number
          nullable: true
      required: [sku]
    InventorySearchResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventorySearchItem"
      required: [data]
    MandateAdjustRequest:
      type: object
      properties:
        sku:
          type: string
        operation:
          type: string
          enum: [add, subtract, set]
        amount:
          type: integer
          minimum: 0
      required: [sku, operation, amount]
    MandateAdjustResponse:
      type: object
      properties:
        sku:
          type: string
        mandateQuantity:
          type: integer
        updated:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              mlItemId:
                type: string
              success:
                type: boolean
              reason:
                type: string
                nullable: true
            required: [productId, mlItemId, success]
      required: [sku, mandateQuantity, updated, failed, results]
    SkuSyncRequest:
      type: object
      properties:
        sku:
          type: string
          description: SKU a sincronizar
        sourceProductId:
          type: string
          nullable: true
          description: ID interno do produto a ser usado como fonte de estoque (opcional)
      required: [sku]
    SkuSyncResponse:
      type: object
      properties:
        synced:
          type: integer
          description: Quantidade de anúncios sincronizados com sucesso (incluindo fonte)
        skipped:
          type: integer
          description: Quantidade de anúncios que falharam na sincronização
        sku:
          type: string
        newStock:
          type: integer
          description: Estoque propagado para todos os anúncios irmãos
      required: [synced, skipped, sku, newStock]

    # --------------------------------------------------------------------------
    # Orders
    # --------------------------------------------------------------------------
...